[Clean Up] Seperate the code to seperate files and call it from a main file
